Add of all changes that make the app synchronized with the server
Tests don't pass yet because of the difficulty to test with the
server. However, the communication with the server works automa-
tically (using the appropriate function in the Firebase class or
the addXp button).

To use the server, simply use the function using the string static
constants in Utils.java for the entries.

The workflow is as follows:
When someone enter the app, before logging in he/she get a
defautl Player instance that is not in communication with the
server. After logging in, if he is new to the app a new section
for his/her variable will be created in the database online with
default values, if not the last state where the Player class was
left with will be re-implemented.

For the moment, the synchronization with the server is simple;
every modification is written to the server.