Require user to login before accessing home screen
In the past the process of logging in was done after the main activity
was loadeed. This doesn't make sense, and logging in should be required
before use of the app.

This commit does the following:
- A new activity, LoginActivity has been added, which displays a button
to login with Tequila.
- Upon clicking this button, the user is sent to Tequila. Afterward,
Tequila redirects to AuthenticationActivity.
- Upon successful authentication the user is sent to MainActivity, and
a login success message is displayed via toast
- Upon failure, the user is sent back to the login activity with a login
failed message