LOGINACTIVITY + CONSTRAINTS + NAVIGATION CLASS UPDATED
Rules added on the login activity. I added a class: ViewPagerAdapter in order to create the image Slider with dots. I updated the constraints of all the layout in order that they fit with all device screens size (I have replaced the Relative Layout by a Constraint Layout). I resolved the Navigation problem. It was not useful to turn it into an interface since this class (which extends AppCompatActivity) is only used by activities and all activities must extend AppCompatActivity(including MapsActivity, so I remove the Map idToAct in Utils.java since it was not used anymore).