Add tests for login and authentication activities
These activities have not yet been tested. Testing for handling of
server responses, for which the logic has been delegated to modular
functions, can be found as unit tests in AuthTest.
The tests in this commit run LoginActivity and AuthenticationActivity,
with the different possibilities of data they would receive in their
intents, to ensure that they do not crash for any of these
possibilities.