Add persist login to the application #105
Now when launching the app, the first thing that is done is check if
the cache contain data of user to log him/her automatically.
If the cache contains it and it has a valid format it will log in the
user in the application (redirected to either mainActivity or
AddQuestionActivity), if not it will behave as before (ask user to
choose role and login).
The path of the data in cache is present in Utils, to read and write to
the cache I use a library from the web (https://github.com/kosalgeek/FileCacher).

For the moment, if you want to change your login, in particular your role,
you just have to comment the first line of the onCreate method in
LoginActivity.