Bug fixes with refactor
This commit fixes two bugs in the refactor.
The first is that firebase data was not being loaded in time when
reading from cache. This was resolved with raising the wait time.
The second is that questions were not loading in AddQuestionActivity for
teachers. This was due to a missing "!" when checking for mock testing.