Implement functionality for enrolling courses
This commit creates an activity, accessible via the settings activity,
for enrolling in courses.
The user may select from a predetermined list of courses to enroll in.
As a teacher, they will be able to add questions for these courses.
As a student, they will be able to view/answer questions for these
courses.