Implement proper workflow for setting courses
With this commit, a user may go to settings, update courses, and select
their courses. When they save their choice, this data is updated in
Firebase. They are redirected to SettingsActivity.

An addition change was made in this commit to the auth process:
Upon logging in, a new global access variable, HOME_ACTIVITY is set, to
be used throughout the app rather than determining it on the fly.