Add function to retrieve courses in Firebase
The function get all the courses that matters at the moment
(depending on variables that will be present shorty (coursesEnrolled
and coursesTeached) and the role).
For now the method call a fictional method in the future activity
ScheduleActivity, that receive a Map<Course, List<Date>> as argument
and adapt the layout.